# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'TaskManagement'
  description: "Task Management è un servizio che permette la gestione e 
l'assegnamento di esami (ovvero gruppi di esercizi) a studenti. Questi 
unltimi dovranno completare gli esercizi. Dopo una deadline definita dal 
T.A. che ha creato l'esame, gli stuenti riceveranno esercizi svolti da 
altri studenti che dovranno correggere. Infine il T.A. darà un voto per 
ogni esercizio e per l'esame ad ogni studente."
host: task-management-solo-numeri-0.herokuapp.com
basePath: /v1
schemes: 
 - https
consumes: 
 - application/json
produces: 
 - application/json
paths:
...
  /users/{id}/submissions:
    get:
      parameters: 
        - in: path
          name: id
          required: true
          type: integer
          minimum: 1
          description: The user ID.
      summary: "Returns user's submission." 
      description: "Returns the submissions made by the specified user." 
      responses:
        200:
          description: "List found"
          schema:
            $ref: "#/definitions/submissionsList"
        404:
          description: Invalid user ID. Bad request.
...
  /submissions:
    post:
      parameters:
        - in: body
          name: submission
          schema:
            $ref: "#/definitions/submission"
      responses:
        201:
          description: new submission created.
          schema:
            $ref: "#/definitions/submissionBody"
        400:
          description: Bad request.
        503:
          description: Service unavailable.
  /submissions/{id}:
    put:
      parameter:
        - in: path
          name: id
          required: true
          type: integer
          description: the submission ID
        - in: body
          name: submission
          schema:
              $ref: "#/definitions/submission"
      responses:
        204:
          description: Submission updated successfully.
        409: 
          description: Conflict. Can't update.
definitions:
...
  submission:
    type: object
    properties:
      id:
        type: integer
      body:
        $ref: "#/definitions/submissionBody"
...
  submissionsList:
    type: array
    items:
      $ref: "#/definitions/submission"
...
  submissionBody:
    type: object
    properties:
      user:
        type: integer
      task:
        type: integer
      exam:
        type: integer
      answer:
        type: string
      final-mark:
        type: integer